Programmed ammo funcionality

+++ Assets/Scripts/PickUps/PickUpStats.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "PickUp Manager/PickUp Stats")]
 public class PickUpStats : ScriptableObject
 {
-    public float pointsToAdd;
+    public int pointsToAdd;
     public PickUpType pickUpType;
 }
